Refactored predicate location and other small changes
