+++ src/main/java/com/minecraftabnormals/abnormals_core/core/AbnormalsCore.java
@@ -8,11 +8,13 @@ import com.minecraftabnormals.abnormals_core.common.world.storage.tracking.DataP
 import com.minecraftabnormals.abnormals_core.common.world.storage.tracking.TrackedData;
 import com.minecraftabnormals.abnormals_core.common.world.storage.tracking.TrackedDataManager;
 import com.minecraftabnormals.abnormals_core.core.api.banner.BannerManager;
-import com.minecraftabnormals.abnormals_core.core.api.conditions.config_predicates.EqualsPredicate;
-import com.minecraftabnormals.abnormals_core.core.api.conditions.config_predicates.GreaterThanOrEqualPredicate;
-import com.minecraftabnormals.abnormals_core.core.api.conditions.config_predicates.GreaterThanPredicate;
-import com.minecraftabnormals.abnormals_core.core.api.conditions.config_predicates.LessThanOrEqualPredicate;
-import com.minecraftabnormals.abnormals_core.core.api.conditions.config_predicates.LessThanPredicate;
+import com.minecraftabnormals.abnormals_core.core.api.conditions.config.ContainsPredicate;
+import com.minecraftabnormals.abnormals_core.core.api.conditions.config.EqualsPredicate;
+import com.minecraftabnormals.abnormals_core.core.api.conditions.config.GreaterThanOrEqualPredicate;
+import com.minecraftabnormals.abnormals_core.core.api.conditions.config.GreaterThanPredicate;
+import com.minecraftabnormals.abnormals_core.core.api.conditions.config.LessThanOrEqualPredicate;
+import com.minecraftabnormals.abnormals_core.core.api.conditions.config.LessThanPredicate;
+import com.minecraftabnormals.abnormals_core.core.api.conditions.config.MatchesPredicate;
 import com.minecraftabnormals.abnormals_core.core.registry.ACEntities;
 import com.minecraftabnormals.abnormals_core.core.registry.ACTileEntities;
 import com.minecraftabnormals.abnormals_core.core.util.DataUtil;
@@ -93,7 +95,8 @@ public final class AbnormalsCore {
 		DataUtil.registerConfigPredicate(new GreaterThanPredicate.Serializer());
 		DataUtil.registerConfigPredicate(new LessThanOrEqualPredicate.Serializer());
 		DataUtil.registerConfigPredicate(new LessThanPredicate.Serializer());
-
+		DataUtil.registerConfigPredicate(new ContainsPredicate.Serializer());
+		DataUtil.registerConfigPredicate(new MatchesPredicate.Serializer());
 
 		BannerManager.RECIPE_SERIALIZERS.register(modEventBus);
