+++ src/main/java/com/minecraftabnormals/abnormals_core/core/api/conditions/config/GreaterThanPredicate.java
@@ -1,4 +1,4 @@
-package com.minecraftabnormals.abnormals_core.core.api.conditions.config_predicates;
+package com.minecraftabnormals.abnormals_core.core.api.conditions.config;
 
 import com.google.gson.JsonObject;
 import com.google.gson.JsonSyntaxException;
